Various kernel functions for MCMC

Various kernel functions for MCMC

kernel_new(proposal, ..., logratio = NULL, kernel_env = new.env(hash = TRUE))

# S3 method for fmcmc_kernel
print(x, ...)

kernel_adapt(
  mu = 0,
  bw = 0L,
  lb = -.Machine$double.xmax,
  ub = .Machine$double.xmax,
  freq = 50L,
  warmup = 500L,
  Sigma = NULL,
  Sd = NULL,
  eps = 1e-04,
  fixed = FALSE
)

kernel_normal(mu = 0, scale = 1, fixed = FALSE, scheme = "joint")

kernel_normal_reflective(
  mu = 0,
  scale = 1,
  lb = -.Machine$double.xmax,
  ub = .Machine$double.xmax,
  fixed = FALSE,
  scheme = "joint"
)

kernel_ram(
  mu = 0,
  eta = function(i, k) min(c(1, i^(-2/3) * k)),
  arate = 0.234,
  freq = 1L,
  warmup = 0L,
  Sigma = NULL,
  eps = 1e-04,
  lb = -.Machine$double.xmax,
  ub = .Machine$double.xmax,
  fixed = FALSE
)

kernel_unif(min. = -1, max. = 1, fixed = FALSE, scheme = "joint")

kernel_unif_reflective(
  min. = -1,
  max. = 1,
  lb = min.,
  ub = max.,
  fixed = FALSE,
  scheme = "joint"
)

Arguments

proposal, logratio
Functions. The function receives a single argument, an environment. This functions are called later within MCMC (see details).

...
In the case of kernel_new, further arguments to be stored with the kernel.

kernel_env
Environment. This will be used as the main container of the kernel's components. It is returned as an object of class c("environment", "fmcmc_kernel").

x
An object of class fmcmc_kernel.

mu, scale
Either a numeric vector or a scalar. Proposal mean and scale.

bw
Integer scalar. The bandwidth, is the number of observations to include in the computation of the variance-covariance matrix.

lb, ub
Either a numeric vector or a scalar. Lower and upper bounds for bounded kernels.

freq
Integer scalar. Frequency of updates. How often the variance-covariance matrix is updated.

warmup
Integer scalar. The number of iterations that the algorithm has to wait before starting to do the updates.

Sigma
The variance-covariance matrix. By default this will be an identity matrix during the warmup period.

Sd
Overall scale for the algorithm. By default, the variance-covariance is scaled to \(2.4^2/d\), with \(d\) the number of dimensions.

eps
Double scalar. Default size of the initial step (see details).

fixed
Logical scalar or vector. When TRUE fixes the corresponding parameter, avoiding new proposals.

scheme
scheme in which proposals are made (see details).

eta
A function that receives the MCMC environment. This is to calculate the scaling factor for the adaptation.

arate
Numeric scalar. Objective acceptance rate.

min., max.
Passed to stats::runif.

Details

The objects fmcmc_kernels are environments that in general contain the following objects:

proposal: The function used to propose changes in the chain based on the current state. The function must return a vector of length equal to the number of parameters in the model.

logratio: This function is called after a new state has been proposed, and is used to compute the log of the Hastings ratio.

In the case that the logratio function is not specified, then it is assumed that the transition kernel is symmetric, this is, log-ratio is then implemented as function(env) {env$f1 - env$f0}

...: Further objects that are used within those functions.

Both functions, proposal and logratio, receive a single argument, an environment, which is passed by the MCMC function during each step using the function environment. The passed environment is actually the environment in which the MCMC function is running, in particular, this environment contains the following objects:

Object		Description

i		Integer. The current iteration.

theta1		Numeric vector. The last proposed state.

theta0		Numeric vector. The current state

f		The log-unnormalized posterior function (a wrapper of fun passed to MCMC).

f1		The last value of f(theta1)

f0		The last value of f(theta0)

kernel		The actual fmcmc_kernel object.

ans		The matrix of samples defined up to i - 1.

These are the core component of the MCMC function. The following block of code is how this is actually implemented in the package:

for (i in 1L:nsteps) {
  # Step 1. Propose
  theta1[] <- kernel$proposal(environment())
  f1       <- f(theta1)
  
  # Checking f(theta1) (it must be a number, can be Inf)
  if (is.nan(f1) | is.na(f1) | is.null(f1)) 
    stop(
      "fun(par) is undefined (", f1, ")",
      "Check either -fun- or the -lb- and -ub- parameters.",
      call. = FALSE
    )
  
  # Step 2. Hastings ratio
  if (R[i] < kernel$logratio(environment())) {
    theta0 <- theta1
    f0     <- f1
  }
  
  # Step 3. Saving the state
  ans[i,] <- theta0
  
}

For more details see the vignette vignette("user-defined-kernels", "fmcmc").

Proposal scheme

The parameter scheme present on the currently available kernels sets the way in which proposals are made. By default, scheme = "joint", proposals are done jointly, this is, at each step of the chain we are proposing new states for each parameter of the model. When scheme = "ordered", a sequential update schema is followed, in which, at each step of the chain, proposals are made one variable at a time, If scheme = "random", proposals are also made one variable at a time but in a random scheme.

Finally, users can specify their own sequence of proposals for the variables by passing a numeric vector to scheme, for example, if the user wants to make sequential proposals following the scheme 2, 1, 3, then scheme must be set to be scheme = c(2, 1, 3).

Creating your own kernels

The function kernel_new is a helper function that allows creating fmcmc_kernel which is used with the MCMC function. The fmcmc_kernel are the backbone of the MCMC function.

Kernels

kernel_adapt Implements the adaptive Metropolis (AM) algorithm of Haario et al. (2001). If the value of bw is greater than zero, then the algorithm folds back AP, a previous version which is known to have ergodicity problems.

The parameter eps has two functions. The first one is to set the initial scale for the multivariate normal kernel, which is replaced after warmup steps with the actual variance-covariance computed by the main algorithm. The second usage is in the equation that ensures that the variance-covariance is greater than zero, this is, the \(\varepsilon\) parameter in the original paper.

The kernel_normal function provides the canonical normal kernel with symmetric transition probabilities.

The kernel_normal_reflective implements the normal kernel with reflective boundaries. Lower and upper bounds are treated using reflecting boundaries, this is, if the proposed \(\theta'\) is greater than the ub, then \(\theta' - ub\) is subtracted from \(ub\). At the same time, if it is less than lb, then \(lb - \theta'\) is added to lb iterating until \(\theta\) is within [lb, ub].

In this case, the transition probability is symmetric (just like the normal kernel).

The kernel_ram Implements Vihola (2012)'s Robust Adaptive Metropolis.

The kernel_unif function provides a uniform transition kernel. This (symmetric) kernel function by default adds the current status values between [-1,1].

The kernel_unif_reflective is similar to kernel_unif with the main difference that proposals are bounded to be within [lb, ub].

References

Brooks, S., Gelman, A., Jones, G. L., & Meng, X. L. (2011). Handbook of Markov Chain Monte Carlo. Handbook of Markov Chain Monte Carlo.

Examples

# Example creating a multivariate normal kernel using the mvtnorm R package
# for a bivariate normal distribution
library(mvtnorm)

# Define your Sigma
sigma <- matrix(c(1, .2, .2, 1), ncol = 2)

# How does it looks like?
sigma

#>      [,1] [,2]
#> [1,]  1.0  0.2
#> [2,]  0.2  1.0

#      [,1] [,2]
# [1,]  1.0  0.2
# [2,]  0.2  1.0

# Create the kernel
kernel_mvn <- kernel_new(
  proposal = function(env) {
  env$theta0 + as.vector(mvtnorm::rmvnorm(1, mean = 0, sigma = sigma.))
  },
  sigma. = sigma
)

# As you can see, in the previous call we passed sigma as it will be used by
# the proposal function
# The logaratio function was not necesary to be passed since this kernel is
# symmetric.