Parameters' update sequence

Parameters' update sequence

plan_update_sequence(k, nsteps, fixed, scheme)

Arguments

k

Integer. Number of parameters

nsteps

Integer. Number of steps.

fixed

Logical scalar or vector of length k. Indicates which parameters will be treated as fixed or not. Single values are recycled.

scheme

Scheme in which the proposals are made (see details).

Value

A logical vector of size nsteps x k.

Details

The parameter scheme present on the currently available kernels sets the way in which proposals are made. By default, scheme = "joint", proposals are done jointly, this is, at each step of the chain we are proposing new states for each parameter of the model. When scheme = "ordered", a sequential update schema is followed, in which, at each step of the chain, proposals are made one variable at a time, If scheme = "random", proposals are also made one variable at a time but in a random scheme.

Finally, users can specify their own sequence of proposals for the variables by passing a numeric vector to scheme, for example, if the user wants to make sequential proposals following the scheme 2, 1, 3, then scheme must be set to be scheme = c(2, 1, 3).